+++ hikari/interactions/command_interactions.py
@@ -50,7 +50,7 @@ from hikari.internal import attr_extensions
 if typing.TYPE_CHECKING:
     from hikari import guilds
     from hikari import permissions as permissions_
-    from hikari import users
+    from hikari import users as users_
     from hikari.api import special_endpoints
 
 
@@ -91,7 +91,7 @@ class InteractionChannel(channels.PartialChannel):
 class ResolvedOptionData:
     """Represents the resolved objects of entities referenced in a command's options."""
 
-    users: typing.Mapping[snowflakes.Snowflake, users.User] = attr.field(repr=False)
+    users: typing.Mapping[snowflakes.Snowflake, users_.User] = attr.field(repr=False)
     """Mapping of snowflake IDs to the resolved option user objects."""
 
     members: typing.Mapping[snowflakes.Snowflake, base_interactions.InteractionMember] = attr.field(repr=False)
@@ -167,7 +167,7 @@ class BaseCommandInteraction(base_interactions.PartialInteraction):
         contains the member's permissions in the current channel.
     """
 
-    user: users.User = attr.field(eq=False, hash=False, repr=True)
+    user: users_.User = attr.field(eq=False, hash=False, repr=True)
     """The user who triggered this command interaction."""
 
     command_id: snowflakes.Snowflake = attr.field(eq=False, hash=False, repr=True)