+++ tests/hikari/impl/test_cache.py
@@ -1502,6 +1502,9 @@ class TestCacheImpl:
             is_deaf=False,
             is_mute=True,
             is_pending=False,
+            raw_communication_disabled_until=datetime.datetime(
+                2021, 10, 18, 13, 11, 18, 384554, tzinfo=datetime.timezone.utc
+            ),
         )
 
         member = cache_impl._build_member(cache_utilities.RefCell(member_data))
@@ -1517,6 +1520,9 @@ class TestCacheImpl:
         assert member.is_deaf is False
         assert member.is_mute is True
         assert member.is_pending is False
+        assert member.raw_communication_disabled_until == datetime.datetime(
+            2021, 10, 18, 13, 11, 18, 384554, tzinfo=datetime.timezone.utc
+        )
 
     def test_clear_members(self, cache_impl):
         mock_user_1 = cache_utilities.RefCell(mock.Mock(id=snowflakes.Snowflake(2123123)))
@@ -1850,6 +1856,9 @@ class TestCacheImpl:
             guild_avatar_hash="gay",
             is_mute=False,
             is_pending=True,
+            raw_communication_disabled_until=datetime.datetime(
+                2021, 10, 18, 13, 11, 18, 384554, tzinfo=datetime.timezone.utc
+            ),
         )
         cache_impl._set_user = mock.Mock(return_value=mock_user_ref)
         cache_impl._increment_ref_count = mock.Mock()
@@ -1879,6 +1888,9 @@ class TestCacheImpl:
         assert member_entry.object.is_deaf is True
         assert member_entry.object.is_mute is False
         assert member_entry.object.is_pending is True
+        assert member_entry.object.raw_communication_disabled_until == datetime.datetime(
+            2021, 10, 18, 13, 11, 18, 384554, tzinfo=datetime.timezone.utc
+        )
 
     def test_set_member_doesnt_increment_user_ref_count_for_pre_cached_member(self, cache_impl):
         mock_user = mock.Mock(users.User, id=snowflakes.Snowflake(645234123))